Add upload picture function
