+++ app/Resources/views/project/index.html.twig
@@ -21,6 +21,7 @@ Project list
             <table class="table table-striped table-hover">
                 <thead>
                     <tr>
+                        <th></th>
                         <th>Name</th>
                         <th>Topic</th>
                         <th>School</th>
@@ -31,6 +32,13 @@ Project list
                 <tbody>
                     {% for project in projects %}
                     <tr>
+                        <td>
+                            {% if project.picture %}
+                            <img class="img-rounded" src="{{ vich_uploader_asset(project.picture, 'imageFile')|imagine_filter('thumbnail_xs') }}" alt="Project picture" />
+                            {% else %}
+                            <img class="img-rounded" src="{{ asset('/template/project/initial_project.png')|imagine_filter('thumbnail_xs') }}" alt="Project picture" />
+                            {% endif %}
+                        </td>
                         <td><a href="{{ path("member_project_show", {"id": project.id}) }}">{{ project.name }}</a></td>
                         <td>{{ project.topic }}</td>
                         <td>{{ project.organization.name }}</td>
