+++ app/Resources/views/user/index.html.twig
@@ -36,9 +36,9 @@ Members
                             <td>{{ loop.index }}</td>
                             <td>
                                 {% if user.picture != null %}
-                                <img class="profile-picture-xs img-circle" src="{{ vich_uploader_asset(user.picture, 'imageFile') }}" alt="{{ user.picture.imageName }}" />
+                                <img class="img-circle" src="{{ vich_uploader_asset(user.picture, 'imageFile')|imagine_filter('thumbnail_xs') }}" alt="Profile picture" />
                                 {% else %}
-                                <img class="profile-picture-xs img-circle" src="{{ asset('template/user/initial_user.png') }}">
+                                <img class="img-circle" src="{{ asset('template/user/initial_user.png')|imagine_filter('thumbnail_xs') }}">
                                 {% endif %}
                             </td>
                             <td><a href="{{ path("member_user_show", {"id": user.id}) }}">{{ user.getFullnamewithTitle() }}</a></td>
