+++ app/Resources/views/user/show.html.twig
@@ -14,9 +14,9 @@ Profile <small>- {{ user.getFullname }}</small>
             {% include "components/member/profile_completeness.html.twig" %}
         {% else %}
             {% if user.picture %}
-            <img class="profile-picture-md" src="{{ vich_uploader_asset(user.picture, 'imageFile') }}" alt="{{ user.picture.imageName }}" />
+            <img class="img-circle" src="{{ vich_uploader_asset(user.picture, 'imageFile')|imagine_filter('thumbnail_sm') }}" alt="User profile picture" />
             {% else %}
-            <img class="profile-picture-md" src="{{ asset('/template/user/initial_user.png') }}" alt="User profile picture" />
+            <img class="img-circle" src="{{ asset('/template/user/initial_user.png')|imagine_filter('thumbnail_sm') }}" alt="User profile picture" />
             {% endif %}
         {% endif %}
         </div>
